zoom-scrape/requirements.txt
flask
requests
waitress
azure-identity
azure-storage-blob
